Fix source root stripping to not strip `requirements.txt` files in plugins (#11712)
The `PythonRequirementsFileSources` is like a `files()` target in that its source roots should not be stripped. However, in https://github.com/pantsbuild/pants/pull/9640, we "fixed" it to not subclass `FilesSources` because this was causing lots of issues with the file being unintentionally included in chroots. Yet, as a result, using `await Get(StrippedSourceFiles)` will try to strip the file, unless the plugin author uses `for_sources_type` and leaves off the type in the allowlist.

This PR allows us to properly handle both use cases by generifying the mechanism for "unrooted sources". Now, any `Sources` subclass can express whether it uses source roots, i.e. it's no longer hardcoded to `FilesSources`.

[ci skip-rust]
[ci skip-build-wheels]