Redirect pantsd stdio to the pants.log rather than entirely closing it. (#11731)
### Problem

`pantsd` has closed the file handles associated with `sys.std*` (fds "0, 1, 2") since its early inception, and until #11536 it allowed exactly one client at a time in order to redirect those precise fd numbers to the client socket (via nailgun).

It's likely that:
1. before #11536, between the closing of stdio and its replacement when a client connected, any attempt to interact with `sys.std*` or Rust's stdio file handles would crash.
2. after #11536, interactions with `sys.std*` or Rust's stdio file handles at _any_ time would crash.

### Solution

Rather than closing `sys.std*` in `pantsd`, replace them with valid file handles as a backstop against code that has-not/cannot be updated to avoid touching them (such as Rust panic handlers). In the case of `stdout`/`stderr`, redirect them to append to the `pants.log`.

### Result

Given what was reported on https://github.com/tokio-rs/tokio/issues/3499, it's likely that this fixes #11364 by replacing the "panicked while processing panic" message with an actual panic message rendered in the `pants.log` (for which we can open a new ticket). 

[ci skip-rust]
[ci skip-build-wheels]