Do not infer dependencies in some edge cases with ambiguity (#11780)
There were some times that we did not correctly handle ambiguous models, specifically when one implementation (e.g. Protobuf) had ambiguity so threw out the module, and another implementation (e.g. 3rd party deps) did not have ambiguity, so it was used. In this scenario, the module should be spoiled for the entire universe: we need to propagate the ambiguity out to other implementations.

[ci skip-rust]