Dependency inference: can disambiguate modules with multiple owners via `!` ignores (#11786)
We are soon adding a warning when dependency inference fails because >1 target exports the same module. For users to silence that warning, they either must explicitly depend on one of the ambiguous modules _or_ ignore some/all of them. Why allow ignores? A user might end up wanting to ignore all of the ambiguous modules so that the module is never included.

To be consistent, because we allow ignoring all ambiguous modules to turn off dep inference for that module, ignoring all _but one_ of the ambiguous targets will disambiguate the inference. (In practice, we expect it to be more likely users will opt-in to the specific ambiguous target they want, rather than opting out of all but one.)

[ci skip-rust]
[ci skip-build-wheels]