Build wheels for the release with GitHub actions (#11752)
This allows us to finish our migration from Travis to GitHub actions.

### Building pants.pex

Note that we do not migrate over the building of pants.pex. To save on CI and S3 resources, we no longer want to deploy unstable PEXes. While unstable wheel files are consumed downstream with the `PANTS_SHA` feature that allows running from any arbitrary commit, that is not wired up to the PEX. The anticipated usage of unstable PEXes is too low to justify the cost.

We do still want to release stable PEXes for now, but to facilitate landing this migration, we will have releases run `build-support/bin/release.sh -p` locally on their machine and then upload the artifact to GitHub releases, similar to the old workflow for Pex releases. This can be automated in a followup.

### macOS version

We now build on macOS 10.15 instead of 10.11. This means that earlier macOS users will not be able to use the latest Pants releases, as announced in https://groups.google.com/u/1/g/pants-devel/c/Bfk4e_9JiT4.