Warn if config files detected but not wired up to Pants (#11852)
A gotcha when setting up Pants is that you must tell Pants about config files, whereas running the underlying tools directly usually auto-detects them. We have a warning at https://www.pantsbuild.org/v2.4/docs/python-linters-and-formatters#configuring-the-tools-and-adding-plugins, but it's really easy to miss that in our docs and a runtime warning reduces the risk of confusion here.

Example warning:

> 12:55:14.71 [WARN] The option `[isort].config` is not configured, but Pants detected relevant config files at ['.isort.cfg', 'pyproject.toml']. Did you mean to set the option `[isort].config`? Pants requires that you explicitly set up config files for them to be used.
> 
> (If you do no want to use this config, you can ignore this warning by adding `ignore_warnings = ["The option `[isort].config` is not configured"]` to `pants.toml` in the `GLOBAL` section.)

To land this, we factor up a `ConfigFilesRequest -> ConfigFiles` rule, which reduces the duplication across our formatters/linters. This will be a good vehicle for us to start doing fancier things with config files, including possibly a) using auto-discovered config files, and b) warning if there are global config files in $HOME.

[ci skip-rust]
[ci skip-build-wheels]