Require `WorkunitsCallback` to set `can_finish_async` (#11687)
Performance-wise, it's important that eligible streaming workunit callbacks are marked that they can finish async, as this avoids the plugin from slowing down Pants's wall-time. However, it is not safe for us to default to this because log statements won't work as expected, which is confusing.

While we could default to not being async compatible, that risks missing out on performance gains that plugin authors aren't aware of. So, we instead require the plugin author to decide what makes sense for their plugin.

[ci skip-rust]
[ci skip-build-wheels]