Deflake tar extraction by setting the cwd through the engine (#11867)
Our tar extraction sometimes flaked as seen in https://github.com/pantsbuild/pants/pull/11856. For an unknown reason, the `output_digest` would not have the leading `__output` prefix. This only has happened in CI and was not reproducible on macOS or an Ubuntu devbox locally. Further, it was flaky.

This new implementation instead uses the engine to set the current working directory, rather than instructing `tar` and `unzip` to do that.

[ci skip-rust]