Show debug level logs in dynamic UI (#11873)
https://github.com/pantsbuild/pants/pull/11816 resulted in the dynamic UI being empty most the time, as the vast majority of our workunits are debug or trace level. 

This system really needs to be redesigned so that log level != dynamic UI setting. There's a tension that we have lots of rules that get run frequently whereby we don't want to log them to avoid flooding the logs, but we do want to show them in the dynamic UI. Related, we want to have dynamic information in the rule's description to make them less vague. This PR punts on both of those due to lack of time right now to revisit the topic.

[ci skip-build-wheels]