Fix remote cache storing of `output_directories` (#11895)
We need to store the `path` in `OutputDirectory`, whereas we had it as the empty string before. This caused `Process.output_directories` to not work properly with the remote cache because we would not restore the `output_directory` prefix when loading from the cache. For example, if the user set `output_directories=('a',)`, which captured `a/b/f.ext`, then we would only store `b/f.ext` instead of `a/b/f.ext`.

Fixes https://github.com/pantsbuild/pants/pull/11856.

[ci skip-build-wheels]