[internal] Port building release Pex to Python (#11958)
More progress on https://github.com/pantsbuild/pants/pull/11952. There is nothing left in Bash except for the processing of options, which then pipes to Python; and setting the Python interpreter to run the Python script. `release.sh` will be removed in a followup.

This adds lightweight validation to our PEX build process, which would have caught https://github.com/pantsbuild/pants/issues/11954.

Certainly, we should be building the PEX via `./pants package` instead of directly via PEX. But this is an incremental improvement from before.

[ci skip-rust]