Fix python process name extraction.
Previous we grabbed the 1st 15 characters on Linux and, at least on some
versions of macOS we didn't grab the process name at all, but the actual
name of the process executable. Since pantsd uses process re-naming to
leave a fingerprint it can use its last recorded pid has not been
recycled to another process, the macOS case is problematic. In out case,
the process executable for pantsd is a Python binary and processes with
that as their executable are ubiquitous. This defeats the pid recycling
detection scheme.

Fix all this by grabbing the process name from the right place, only
falling back to the prior if there is a problem reading the process
command line (can happen for zombie processes at least). Add an
integration test that verifies we now read the full custom command line
and not a truncated version of it or the name of the process executable.

[ci skip-build-wheels]
[ci skip-rust]