Stricter PyPI role checks. (#11999)
Previously we did loose checks just to verify that the deploy was likely to succeed,
and that the deployed packages existed and had a minimum number of owners.

There were a few problems with this:

We required all users to be owners, even though it's fine for them to be maintainers.
We had no way to detect unexpected laxity. As long as the deploying user was an owner,
and there were enough other owners, we would not check who those others were.
In other words, we suffered from the "Jurassic Park oversight": in the novel, the park control system
stops counting dinosaurs when it reaches the expected number, so it fails to detect the supposedly
impossible dinosaur births. Unfortunately this scene didn't make it into the movie, but the lesson is
to look for the unexpected, not just verify the expected.

This change provides much stricter checks. For any PyPI package we can assert that the maintainers
and owners are exactly who we expect.

For deploys we check just the deployed packages.
We replace the "list owners" command with a "validate roles" command.
So instead of listing owners we just assert that they are exactly who we expect.

Note that this change enumerates the mantainers and owners of pantsbuild packages, but this info
is public on PyPI anyway, so we are not revealing anything new.

This change also fixes an ongoing throttling issue with the PyPI XML-RPC API.

[ci skip-rust]
[ci skip-build-wheels]