Fix `--remote-auth-plugin` and `--remote-oauth-bearer-token-path` to execute every run with Pantsd (#12020)
As found in https://github.com/pantsbuild/pants/issues/12018, `--remote-auth-plugin` and `--remote-oauth-bearer-token-path` are both impure functions with side effects like reading from the file system. Consequently, it's necessary to rerun the two mechanisms every Pants run, even with pantsd. For example, currently, we will not pick up changes made to the oauth file unless you kill Pantsd.

This PR fixes it so that we always do the correct things of re-evaluating the `DynamicRemoteExecutionOptions`. If those have changed from the prior Pantsd run, we re-initialize the scheduler.

Naively, however, this has a negative implication for the Toolchain plugin. Because Toolchain's auth logic is stateless, it generates a new short-lived token every call, which updates `--remote-store-header`. That means that this PR will cause Pantsd to be invalidated every single run, making Pantsd useless. This will be fixed in a followup by passing the prior `AuthPluginResult` to the auth plugin function + allowing the plugin to set a token expiration time, which will allow the plugin to decide if it's safe to reuse the prior token.
# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]