move `side_effecting` to its own module
Mypy cannot determine the type of `side_effecting`
when running typecheck on a single file. This issue
has been introduced in #11936 but it seems to be caused
by how mypy resolve imports and assign type at the top-level.

Moving the decorator to its own module, still inside rules, helps
mypy to resolve to the correct type.

[ci skip-rust]