Auth plugins can inspect the prior result, set an expiration, and override `--remote-{store,execution}-address` (#12029)
https://github.com/pantsbuild/pants/pull/12020 results in auth plugins being executed every run of `./pants`, even when Pantsd is used. 

While correct, that is a problem for auth plugins like Toolchain's that are stateless and give back a new token every time. Getting a new token changes the `store_headers`, which changes the bootstrap options sent over FFI to Rust to create the scheduler; we must reinitialize the scheduler in this case, which is slow.

Instead, we now pass `prior_result: AuthPluginResult | None` to the auth plugin hook, which allows plugins to recycle prior results if they'd like. We also allow plugins to set `expiration` on the `AuthPluginResult`, which is particularly useful for plugins deciding to recycle a token vs. generate a new one. 

Finally, we allow the plugin to override `--remote-store-address` and `--remote-execution-address`, which isn't totally related to this PR but has been requested as a feature and is done here to reduce churn in the API.

Combined, this PR allows plugin authors to avoid the problem of invalidating Pantsd every run, while avoiding the problem that Pantsd blocked every regenerating tokens until a restart. Closes https://github.com/pantsbuild/pants/issues/12018.

[ci skip-rust]
[ci skip-build-wheels]