Add plugin hook for setting up Pytest context (#12091)
This new hook allows users to do things like:

- Validate that databases are running.
- Insert certain files, e.g. our `runtime_package_dependencies` feature.

In the future, we can expand it to have new capabilities like setting environment variables.

Unlike our `setup_py` kwargs plugin hook, we do allow you to run >1 plugin on the same target here. It should be safe to merge multiple plugins, and this is a useful feature.

This plugin is only implemented for Pytest, not for other test runners like `shunit2`. It seems unlikely a user would want the same setup for Pytest as other languages; and if they do, they can simply factor up a helper rule that gets used by both plugin hooks.

[ci skip-rust]
[ci skip-build-wheels]