Replace `ProcessCacheScope.PER_RESTART` with `ProcessCacheScope.PER_RESTART_ALWAYS` and `ProcessCacheScope.PER_RESTART_SUCCESSFUL` (#12094)
Previously, `ProcessCacheScope.PER_RESTART` always memoized failures (but not cached to disk). 

That wasn't a problem in pantsbuild/pants because we always use `await Get(ProcessResult, Process)` instead of `Get(FallibleProcessResult, Process)`, and the error restarts Pantsd. 

But, it's valid to use a fallible process result and not want to memoize failures. For example, a user is writing a plugin to check if a database is alive before test runs. They must not cache this to disk because it's not safe to, but they also don't want to have to run the check every single distinct test run.

[ci skip-build-wheels]