Allow overriding option default help display. (#12128)
In rare cases we may want the help to show some string
explaining the default value, instead of showing the actual value.

For example, a more instructive help would explain that the default
for --process-execution-local-parallelism is "#cores" and not
whatever the number of cores is on the machine generating the help
(which in the case of the docsite is the releaser's machine, which
is really not useful to the reader).

This change supports this functionality. It is intended to be used
sparingly.  But another place where this could be employed is when
int values take on units (e.g., "512MiB"). We'd want to display
the default as a string including the units, and not as the
underlying raw int (e.g, 537395200).

[ci skip-rust]

[ci skip-build-wheels]