Improve handling of linter reports. (#12122)
Previously we tried to directly control linter report file output options, which led to several problems:
- We hard-coded report file names, including a .txt suffix, which doesn't play well with linters like Bandit that have several possible output formats.
- We assumed reports were a single file.
- We only supported a single report per linter, even if there were partitions.

This change modifies how we handle linter report files, bringing it inline with how we handle PyTest plugin custom output.
- Now all the user has to do to get a report is request it using the linter's standard args or config, and the only constraint is that those must go into a reldir named `reports/`, so that we know to capture it out of the sandbox.
- We capture whatever is in that dir, no matter how many files.
- We output whatever reports we find to dist/ , under a directory that is per-partition, so multiple lint runs can emit reports.

The only caveat is that it's not obvious to the user that they must configure their reports must go under `reports/`. So we must document this clearly. In the future we could potentially attempt to detect this case and warn by heuristically analyzing the args and/or config. But that also might be overkill so we don't do it yet.

[ci skip-rust]

[ci skip-build-wheels]