Allow setting `--pantsd-max-memory-usage` with `GiB`, `MiB`, and `KiB` units (#12123)
Given that we've found this option can be crucial to avoiding the OOM killer, we want to make it more ergonomic to work with: https://github.com/pantsbuild/pants/pull/12107. 

Most users don't think in terms of bytes, but rather MB/GB - we should save them from having to pull out online calculators.

We use the library Pint to robustly support memory size conversions, including to properly support floats without floating-point arithmetic. A major benefit of Pint is that it has a function `.to_compact()` that allows us to generate a meaningful representation of the number of bytes, e.g. knowing when to say `12MB` vs `1GB` based on how many bytes there are. Pint has no dependencies, has a wheel, and is actively maintained.

[ci skip-build-wheels]