Hash the repo_id in the tag.
Previously we sent the unhashed id in the tag, but the hashed
id as a standalone field.

This is only a minor anonymity issue, since the raw repo_id
is just a UUID. But it is a uniformity issue, and it violates
our promise to users.

Separately I will have the existing data processed to replace
the offending tags.

[ci skip-rust]

[ci skip-build-wheels]