Temporarily fix a test that relies on the state of the world. (#12153)
certifi published a new version, and this broke some tests of
pex resolution that assumed an earlier version would be resolved.

This is a short-term fix to get tests running, but this test
needs to be fixed to not depend on external state.

[ci skip-rust]

[ci skip-build-wheels]

(cherry picked from commit 2d62aa954145a071bf0536ca775f7f9aac0e2f0d)