Fix `Key` interning race. (#12152)
Previously a `Key` could be observed before it was added to the
`reverse_keys` mapping leading to a panic in `key_get`. That panic used
a debug format on the problematic key and the `Debug` impl for `Key`
indirectly uses `key_get`. This results in a panic while handling a
panic.

Fix the root cause of the panic, ensuring the panic can never happen
as was intended. Also fix the recursive use of `key_get` in the panic
for sanity sake.

Fixes #11926

(cherry picked from commit a59be85e10fc97d2f5774ceca583741da3f0547d)

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]