[internal] Use long-form arguments for `release.sh` (#12017)
Closes https://github.com/pantsbuild/pants/issues/5551.

Originally this PR tried to entirely delete `release.sh`, but it proved difficult to correctly set the Python interpreter to run with, which is necessary for our wheels to be built correctly. Note that the Python version used to invoke `./pants run` should be the same as what the script uses - it's too late to try changing inside the Python script.

This PR instead removes most the logic from `release.sh` and passes through all args to the Python script.

This also removes the `check_shell.sh` script now that we use Shellcheck and we have so little Bash left, and that Shellcheck's SC2155 already covers one of the two checks.