Fallback to `[python-setup].interpreter_constraints` for `python_distribution`s with no first-party dependencies (#12207)
We use interpreter constraints with `python_distribution` targets to determine what to run `setuptools` with if `setup_py_commands` is set, along with determining if Py2 is used for how to handle namespace packages. 

Currently, the ICs can only be set via dependencies on `python_library` targets. If there are no relevant deps—e.g. an empty `python_distribution`, which is weird but legal—we use no constraints at all. This resulted in a bug where Python 2 was being used to build `setuptools>50`, which fails to build.

We fix this by simply falling back to `[python-setup].interpreter_constraints`, which is what we also do with `./pants repl` when there are no relevant targets specified. 

It's possible, but unlikely, that a user won't want to couple `[python-setup].interpreter_constraints` with setuptools, e.g. that their constraints are Py2 or Py3, but they want to use setuptools 50. We decided to not proactively plan for this because it seems unlikely—especially because it is so rare to have a distribution with no first-party deps—and we want to avoid options fatigue. If it becomes a problem, we could add an option `[setuptools].fallback_interpreter_constraints`.

### Rejected alternative: add `interpreter_constraints` to `python_distribution`

We have `interpreter_constraints` on `pex_binary` to accommodate a PEX with only third party dependencies. There, constraints are relevant.

But for `python_distribution`, if you have a dist with no first-party dependencies, the only reason the ICs matter is so that they're compatible with `[setuptools].version`, as it's only used to run setuptools and the ICs are not embedded into the distribution. If you do have first-party deps, the ICs should be updated on the relevant `python_library` code.

[ci skip-rust]