[1.25.x-twtr][jvm] javac --release support, ensure dist selection reflects target platform (#12104)
This patch is for the 1.25.x-twtr branch to support our jdk 11 migration. It fixes an issue where jvm distributions were not selected dependent on the target platform, which meant that pants would fail to compile with 11 for an 11 requesting target if there were also targets using 8.

Additionally, JDK 9+'s javac supports `--release` which allows compilation against earlier JDK APIs from javacs for later ones. If a later JDK is being used, with this patch we'll use --release. In zinc, we'll skip adding -source -target etc if they are present in the platform's compiler args.