[internal] Add mechanism to publish an Apple Silicon wheel (#12227)
Until [GitHub Actions finishes adding support for Apple Silicon](https://github.com/actions/virtual-environments/issues/2187), building wheels for Apple Silicon is tricky. We could use cross-compilation, but it's non-trivial and we haven't had the time to prioritize it.

In the meantime, users have had to use Rosetta, which is subpar and has worse performance. We want Pants to work out of the box.

Instead, two Pants developers now have M1s, myself and @chrisjrn. Rather than using CI, we can build a macos_arm wheel locally and publish to PyPI. We'd explain to users that this only best-effort, e.g. it may take a day or two to get the wheel published. Even though it's not super sustainable to rely on a person to build the wheel—rather than automating—this is better than the status quo of no out-of-the-box support. Once GitHub actions adds support, we'll switch to that.

[ci skip-rust]