Support building dists from preexisting setup.py. (#12250)
Previously we insisted on generating the setup.py from information known to Pants.
This change allows users to build against an existing setup.py checked into the repo.

This is useful when the setup logic is unusual, or involves building native code, and so on.
Of course in this case there may be redundancy between the information known to Pants
and whatever is in the setup.py. So the author of the setup.py is responsible for ensuring that
it is correct with respect to versions, dependencies and so on.

This will make it easier to adopt Pants in an existing repo with existing setup.py files.

To use this new feature, your provides=setup_py(...) stanza should contain a setup_script key,
pointing to the setup script, and it should also contain the name and version, repeated from
the setup.py. It may contain no other keys.

[ci skip-rust]

[ci skip-build-wheels]