[internal] `./pants lock` uses transitive closure as inputs (#12312)
These are _not_ the proposed final semantics for the `./pants lock` goal. The semantics will change when we have per-tool lockfiles and multiple lockfiles, likely such that `./pants lock ::` will regenerate all distinct lockfiles.

For now, this PR merely aligns `./pants lock` with the current `generate_lockfile.sh` script, so that we can run `./pants --tag=-lockfile_ignore lock ::`. The goal is to replace `generate_lockfile.sh` with this goal.

[ci skip-rust]
[ci skip-build-wheels]